OMAP3:CAMERA: Fix for the memory leak in camera.
The member of ofh allocated during the vout buff prepare that was never freed
resulting in memory leak. ofh->vbq released in vout buff release.

The two member of file handle structure are not released, resulting in memory leak.

The above two cases has been taken care in this patch.

Change-Id: I4f95a17dfdf7e0d2a90e5bc7394157289f6ebbef
Signed-off-by: bharath <x0172602@ti.com>